Moved coordinator into module coordinator.py
Imports are already pruned for new code, so there are missing import in this file!
