+++ custom_components/pun_sensor/sensor.py
@@ -31,9 +31,13 @@ from .const import (
 
 ATTR_ROUNDED_DECIMALS = "rounded_decimals"
 
-async def async_setup_entry(hass: HomeAssistant, config: ConfigEntry,
+
+async def async_setup_entry(
+    hass: HomeAssistant,
+    config: ConfigEntry,
     async_add_entities: AddEntitiesCallback,
-    discovery_info: DiscoveryInfoType | None = None) -> None:
+    discovery_info: DiscoveryInfoType | None = None,
+) -> None:
     """Inizializza e crea i sensori"""
 
     # Restituisce il coordinator
@@ -41,7 +45,9 @@ async def async_setup_entry(hass: HomeAssistant, config: ConfigEntry,
 
     # Verifica la versione di Home Assistant
     global has_suggested_display_precision
-    has_suggested_display_precision = (AwesomeVersion(HA_VERSION) >= AwesomeVersion("2023.3.0"))
+    has_suggested_display_precision = AwesomeVersion(HA_VERSION) >= AwesomeVersion(
+        "2023.3.0"
+    )
 
     # Crea i sensori (legati al coordinator)
     entities = []
@@ -56,7 +62,7 @@ async def async_setup_entry(hass: HomeAssistant, config: ConfigEntry,
     # Aggiunge i sensori ma non aggiorna automaticamente via web
     # per lasciare il tempo ad Home Assistant di avviarsi
     async_add_entities(entities, update_before_add=False)
-    
+
 
 def fmt_float(num: float) -> str:
     """Formatta adeguatamente il numero decimale"""
@@ -66,7 +72,8 @@ def fmt_float(num: float) -> str:
     # In versioni precedenti di Home Assistant che non supportano
     # l'attributo 'suggested_display_precision' restituisce il numero
     # decimale già adeguatamente formattato come stringa
-    return format(round(num, 6), '.6f')
+    return format(round(num, 6), ".6f")
+
 
 class PUNSensorEntity(CoordinatorEntity, SensorEntity, RestoreEntity):
     """Sensore PUN relativo al prezzo medio mensile per fasce"""
@@ -79,16 +86,16 @@ class PUNSensorEntity(CoordinatorEntity, SensorEntity, RestoreEntity):
         self.tipo = tipo
 
         # ID univoco sensore basato su un nome fisso
-        if (self.tipo == PUN_FASCIA_F3):
-            self.entity_id = ENTITY_ID_FORMAT.format('pun_fascia_f3')
-        elif (self.tipo == PUN_FASCIA_F2):
-            self.entity_id = ENTITY_ID_FORMAT.format('pun_fascia_f2')
-        elif (self.tipo == PUN_FASCIA_F1):
-            self.entity_id = ENTITY_ID_FORMAT.format('pun_fascia_f1')
-        elif (self.tipo == PUN_FASCIA_MONO):
-            self.entity_id = ENTITY_ID_FORMAT.format('pun_mono_orario')
-        elif (self.tipo == PUN_FASCIA_F23):
-            self.entity_id = ENTITY_ID_FORMAT.format('pun_fascia_f23')
+        if self.tipo == PUN_FASCIA_F3:
+            self.entity_id = ENTITY_ID_FORMAT.format("pun_fascia_f3")
+        elif self.tipo == PUN_FASCIA_F2:
+            self.entity_id = ENTITY_ID_FORMAT.format("pun_fascia_f2")
+        elif self.tipo == PUN_FASCIA_F1:
+            self.entity_id = ENTITY_ID_FORMAT.format("pun_fascia_f1")
+        elif self.tipo == PUN_FASCIA_MONO:
+            self.entity_id = ENTITY_ID_FORMAT.format("pun_mono_orario")
+        elif self.tipo == PUN_FASCIA_F23:
+            self.entity_id = ENTITY_ID_FORMAT.format("pun_fascia_f23")
         else:
             self.entity_id = "None"
         self._attr_unique_id = self.entity_id
@@ -104,23 +111,24 @@ class PUNSensorEntity(CoordinatorEntity, SensorEntity, RestoreEntity):
     def _handle_coordinator_update(self) -> None:
         """Gestisce l'aggiornamento dei dati dal coordinator"""
         self._available = self.coordinator.orari[self.tipo] > 0
-        if (self._available): self._native_value = self.coordinator.pun[self.tipo]
+        if self._available:
+            self._native_value = self.coordinator.pun[self.tipo]
         self.async_write_ha_state()
 
     @property
     def extra_restore_state_data(self) -> ExtraStoredData:
         """Determina i dati da salvare per il ripristino successivo"""
-        return RestoredExtraData(dict(
-            native_value = self._native_value if self._available else None
-        ))
-    
+        return RestoredExtraData(
+            dict(native_value=self._native_value if self._available else None)
+        )
+
     async def async_added_to_hass(self) -> None:
         """Entità aggiunta ad Home Assistant"""
         await super().async_added_to_hass()
 
-        # Recupera lo stato precedente, se esiste        
+        # Recupera lo stato precedente, se esiste
         if (old_data := await self.async_get_last_extra_data()) is not None:
-            if (old_native_value := old_data.as_dict().get('native_value')) is not None:
+            if (old_native_value := old_data.as_dict().get("native_value")) is not None:
                 self._available = True
                 self._native_value = old_native_value
 
@@ -156,15 +164,15 @@ class PUNSensorEntity(CoordinatorEntity, SensorEntity, RestoreEntity):
     @property
     def name(self) -> str:
         """Restituisce il nome del sensore"""
-        if (self.tipo == PUN_FASCIA_F3):
+        if self.tipo == PUN_FASCIA_F3:
             return "PUN fascia F3"
-        elif (self.tipo == PUN_FASCIA_F2):
+        elif self.tipo == PUN_FASCIA_F2:
             return "PUN fascia F2"
-        elif (self.tipo == PUN_FASCIA_F1):
+        elif self.tipo == PUN_FASCIA_F1:
             return "PUN fascia F1"
-        elif (self.tipo == PUN_FASCIA_MONO):
+        elif self.tipo == PUN_FASCIA_MONO:
             return "PUN mono-orario"
-        elif (self.tipo == PUN_FASCIA_F23):
+        elif self.tipo == PUN_FASCIA_F23:
             return "PUN fascia F23"
         else:
             return "None"
@@ -178,9 +186,10 @@ class PUNSensorEntity(CoordinatorEntity, SensorEntity, RestoreEntity):
         # Nelle versioni precedenti di Home Assistant
         # restituisce un valore arrotondato come attributo
         state_attr = {
-            ATTR_ROUNDED_DECIMALS: str(format(round(self.native_value, 3), '.3f'))
+            ATTR_ROUNDED_DECIMALS: str(format(round(self.native_value, 3), ".3f"))
         }
         return state_attr
+
 
 class FasciaPUNSensorEntity(CoordinatorEntity, SensorEntity):
     """Sensore che rappresenta la fascia PUN corrente"""
@@ -192,7 +201,7 @@ class FasciaPUNSensorEntity(CoordinatorEntity, SensorEntity):
         self.coordinator = coordinator
 
         # ID univoco sensore basato su un nome fisso
-        self.entity_id = ENTITY_ID_FORMAT.format('pun_fascia_corrente')
+        self.entity_id = ENTITY_ID_FORMAT.format("pun_fascia_corrente")
         self._attr_unique_id = self.entity_id
         self._attr_has_entity_name = True
 
@@ -209,15 +218,15 @@ class FasciaPUNSensorEntity(CoordinatorEntity, SensorEntity):
     def available(self) -> bool:
         """Determina se il valore è disponibile"""
         return self.coordinator.fascia_corrente is not None
-    
+
     @property
     def state(self) -> str:
         """Restituisce la fascia corrente come stato"""
-        if (self.coordinator.fascia_corrente == 3):
+        if self.coordinator.fascia_corrente == 3:
             return "F3"
-        elif (self.coordinator.fascia_corrente == 2):
+        elif self.coordinator.fascia_corrente == 2:
             return "F2"
-        elif (self.coordinator.fascia_corrente == 1):
+        elif self.coordinator.fascia_corrente == 1:
             return "F1"
         else:
             return "None"
@@ -232,6 +241,7 @@ class FasciaPUNSensorEntity(CoordinatorEntity, SensorEntity):
         """Restituisce il nome del sensore"""
         return "Fascia corrente"
 
+
 class PrezzoFasciaPUNSensorEntity(FasciaPUNSensorEntity, RestoreEntity):
     """Sensore che rappresenta il prezzo PUN della fascia corrente"""
 
@@ -239,7 +249,7 @@ class PrezzoFasciaPUNSensorEntity(FasciaPUNSensorEntity, RestoreEntity):
         super().__init__(coordinator)
 
         # ID univoco sensore basato su un nome fisso
-        self.entity_id = ENTITY_ID_FORMAT.format('pun_prezzo_fascia_corrente')
+        self.entity_id = ENTITY_ID_FORMAT.format("pun_prezzo_fascia_corrente")
         self._attr_unique_id = self.entity_id
         self._attr_has_entity_name = True
 
@@ -254,15 +264,15 @@ class PrezzoFasciaPUNSensorEntity(FasciaPUNSensorEntity, RestoreEntity):
     def _handle_coordinator_update(self) -> None:
         """Gestisce l'aggiornamento dei dati dal coordinator"""
         if super().available:
-            if (self.coordinator.fascia_corrente == 3):
+            if self.coordinator.fascia_corrente == 3:
                 self._available = self.coordinator.orari[PUN_FASCIA_F3] > 0
                 self._native_value = self.coordinator.pun[PUN_FASCIA_F3]
                 self._friendly_name = "Prezzo fascia corrente (F3)"
-            elif (self.coordinator.fascia_corrente == 2):
+            elif self.coordinator.fascia_corrente == 2:
                 self._available = self.coordinator.orari[PUN_FASCIA_F2] > 0
                 self._native_value = self.coordinator.pun[PUN_FASCIA_F2]
                 self._friendly_name = "Prezzo fascia corrente (F2)"
-            elif (self.coordinator.fascia_corrente == 1):
+            elif self.coordinator.fascia_corrente == 1:
                 self._available = self.coordinator.orari[PUN_FASCIA_F1] > 0
                 self._native_value = self.coordinator.pun[PUN_FASCIA_F1]
                 self._friendly_name = "Prezzo fascia corrente (F1)"
@@ -279,21 +289,25 @@ class PrezzoFasciaPUNSensorEntity(FasciaPUNSensorEntity, RestoreEntity):
     @property
     def extra_restore_state_data(self) -> ExtraStoredData:
         """Determina i dati da salvare per il ripristino successivo"""
-        return RestoredExtraData(dict(
-            native_value = self._native_value if self._available else None,
-            friendly_name = self._friendly_name if self._available else None
-        ))
-    
+        return RestoredExtraData(
+            dict(
+                native_value=self._native_value if self._available else None,
+                friendly_name=self._friendly_name if self._available else None,
+            )
+        )
+
     async def async_added_to_hass(self) -> None:
         """Entità aggiunta ad Home Assistant"""
         await super().async_added_to_hass()
 
-        # Recupera lo stato precedente, se esiste        
+        # Recupera lo stato precedente, se esiste
         if (old_data := await self.async_get_last_extra_data()) is not None:
-            if (old_native_value := old_data.as_dict().get('native_value')) is not None:
+            if (old_native_value := old_data.as_dict().get("native_value")) is not None:
                 self._available = True
                 self._native_value = old_native_value
-            if (old_friendly_name := old_data.as_dict().get('friendly_name')) is not None:
+            if (
+                old_friendly_name := old_data.as_dict().get("friendly_name")
+            ) is not None:
                 self._friendly_name = old_friendly_name
 
     @property
@@ -334,6 +348,6 @@ class PrezzoFasciaPUNSensorEntity(FasciaPUNSensorEntity, RestoreEntity):
         # Nelle versioni precedenti di Home Assistant
         # restituisce un valore arrotondato come attributo
         state_attr = {
-            ATTR_ROUNDED_DECIMALS: str(format(round(self.native_value, 3), '.3f'))
+            ATTR_ROUNDED_DECIMALS: str(format(round(self.native_value, 3), ".3f"))
         }
         return state_attr
